--- src/main/java/nextstep/ladder/BooleanGenerator.java
@@ -1,6 +1,0 @@
-package nextstep.ladder;
-
-@FunctionalInterface
-interface BooleanGenerator {
-    Boolean nextBoolean();
-}
