+++ src/main/java/nextstep/ladder/domain/BooleanGenerator.java
@@ -1,0 +1,6 @@
+package nextstep.ladder.domain;
+
+@FunctionalInterface
+public interface BooleanGenerator {
+    Boolean nextBoolean();
+}
