+++ src/main/java/nextstep/ladder/domain/Line.java
@@ -1,9 +1,9 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.List;
 import java.util.stream.Stream;
 
-class Line {
+public class Line {
 
     private final List<Boolean> spoke;
